fix(go2cs): emit plain UL for >MaxInt64 literals in uint64 contexts
Absorbed from chip branch claude/sad-elion-4da4ac (594069617), converter
fix and docs entry only - that branch predates the merged WP-1 work and
its remaining content (its own MathFloatBits project, core/math bit
functions, slnx registration, test-class blocks) is superseded and
deliberately dropped.

Root cause: the ParseUint arm of convBasicLit's INT emission has
UNCONDITIONALLY prefixed `(nuint)` on every literal above MaxUint32
since a771782d4. That bridge is only needed when the literal's resolved
type is Go uint/uintptr (C# nuint - a bare ulong literal has no implicit
conversion to it, CS0266; the non-constant unchecked (nuint) conversion
compiles). In a uint64 context it is spurious: semantically wrong for
the 64-bit target type and value-truncating on a 32-bit platform -
math.Float64frombits(0xFFF0000000000000) emitted
((nuint)0xFFF0000000000000UL) while the int64-range 0x7FF0000000000000
emitted clean (the signed branch already consulted the resolved type via
isUnsignedType). The branch now consults the literal's resolved
UNDERLYING type: uint64 - incl. named types over uint64, whose [GoType]
wrappers convert implicitly from ulong - takes the plain UL literal;
native-width unsigned targets keep the (nuint) cast (LargeUintptrConst
pins that path, byte-identical in check-no-regression).

Corpus drift: exactly three generated files, all nuint-cast-removal
shape - BitwiseUntypedConst (two copysign call args),
NamedIntSignednessConv (timestamp big = 18446744073709551615UL),
ShiftPrecedenceUnsigned (uint64 hi = 0xFFFFFFFFFFFFFFFFUL); goldens
re-baselined via re-transpile + UpdateTestTargets --createTargetFiles,
all three plus LargeUintptrConst green across all 4 phases. The merged
MathFloatBits golden ((nuint)0xFFF0000000000000UL on master) becomes the
fix's pin at the slot-4 rebase re-baseline.

Also folds in the review's cosmetic doc nit: the Reference tightening
prose wrote (uint8) where the golden emits (byte).

+++ src/Tests/Behavioral/BitwiseUntypedConst/main.cs
@@ -10,8 +10,8 @@ internal static uint64 copysign(uint64 f, uint64 sign) {
 }
 
 internal static void Main() {
-    fmt.Println(copysign(0xFF, (nuint)0x8000000000000000UL));
-    fmt.Println(copysign((nuint)0x8000000000000042UL, 0));
+    fmt.Println(copysign(0xFF, 0x8000000000000000UL));
+    fmt.Println(copysign(0x8000000000000042UL, 0));
 }
 
 } // end main_package
